Remove seatd socket when entering and exiting the container
As we don't use `--force-recreate` in our `docker-compose up` command,
we need to manually clean up the socket file which will be invalidated
between consecutive runs of this container, due to `seatd` not unlinking
existing sockets by default.

We don't introduce `--force-recreate` because it may cause in a pile-up
of containers if there are too many restarts.

Related-to: TOR-2973

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>