stress-tests: Improve stressing
*hackbench* doesn't produce much load on a SMP system. A better way to stress
the CPU is copying data from /dev/zero to /dev/null.

Running *du* and *find* doesn't stress Linux very much. Most of it is
optimised through caches. Better: run *dd* to copy the whole block device to
/dev/null.

To stress the network interface (but not your internet connection) send
massive pings to the gateway. In a docker container this might be the system
itself which wouldn't load the NIC. With the environment variable
STRESS_SERVER you can set a different address. The tool *ip* used to find
the gateway is in the package iproute2.

Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>