gitlab-ci: disable pushing images outside DEFAULT_BRANCH branch
DEFAULT_BRANCH stores the name of the branch from which to push
container images to dockerhub. If the ci is launched from another
branch, the images will be built but won't be pushed.

PUSH_TO_REGISTRY is the variable that stores the command argument to
tell buildx wether to push or not to dockerhub.

Related-to: TCCP-20
Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>