am62: weston: entry.sh: launch weston as torizon user
Debian's weston isn't built with libseat support, but we need it our
containers to allow seat management via libseat and also to run wayland
rootless. We added a custom build of weston in our am62's feed with
libseat support, so we can enable the init function in entry.sh to
launch weston as torizon instead of root.

Related-to: TCCP-614
Signed-off-by: Carlos Henrique Lima Melara <carlos.melara@toradex.com>